refactor(manifest): move network manifest schema to shared package
