+++ pkg/plugin/implementation/opapolicychecker/enforcer_test.go
@@ -89,37 +89,6 @@ func validManifestGovernanceYAML() string {
 	)
 }
 
-func validNetworkManifestForTest(networkID string, now time.Time) networkManifest {
-	signed := true
-	return networkManifest{
-		ManifestVersion: "1.0",
-		ManifestType:    "network-manifest",
-		NetworkID:       networkID,
-		ReleaseID:       "2026.02",
-		Publisher: networkManifestPublisher{
-			Role:   "NFO",
-			Domain: "nfh.global",
-		},
-		Policies: &networkManifestPolicies{
-			Type:   "rego",
-			Source: "file",
-			File: &networkManifestFile{
-				ID:                        "network-policy-file",
-				URL:                       "https://example.com/policy.rego",
-				PolicyQueryPath:           "data.logistics.result",
-				Signed:                    true,
-				SignatureURL:              "https://example.com/policy.rego.sig",
-				SigningPublicKeyLookupURL: "https://example.com/public-key",
-			},
-		},
-		Governance: networkManifestGovernance{
-			EffectiveFrom:  now.Add(-1 * time.Hour).Format(time.RFC3339),
-			EffectiveUntil: now.Add(1 * time.Hour).Format(time.RFC3339),
-			Signed:         &signed,
-		},
-	}
-}
-
 func indentYAML(s, prefix string) string {
 	lines := strings.Split(strings.TrimSuffix(s, "\n"), "\n")
 	for i, line := range lines {
@@ -249,108 +218,6 @@ func TestParsePolicyConfig_ManifestType(t *testing.T) {
 	}
 	if cfg.Location != "" || cfg.Query != "" || len(cfg.PolicyPaths) != 0 {
 		t.Fatalf("expected manifest config to defer location/query resolution, got %#v", cfg)
-	}
-}
-
-func TestValidateNetworkManifest(t *testing.T) {
-	now := time.Date(2026, time.April, 22, 12, 0, 0, 0, time.UTC)
-	expectedNetworkID := "nfh.global/testnet"
-
-	tests := []struct {
-		name       string
-		mutate     func(*networkManifest)
-		wantErrSub string
-	}{
-		{
-			name: "valid",
-		},
-		{
-			name: "invalid manifest type",
-			mutate: func(manifest *networkManifest) {
-				manifest.ManifestType = "other"
-			},
-			wantErrSub: `must have manifest_type="network-manifest"`,
-		},
-		{
-			name: "missing policies section",
-			mutate: func(manifest *networkManifest) {
-				manifest.Policies = nil
-			},
-			wantErrSub: "missing policies section",
-		},
-		{
-			name: "network mismatch",
-			mutate: func(manifest *networkManifest) {
-				manifest.NetworkID = "example/logistics"
-			},
-			wantErrSub: `does not match configured network "nfh.global/testnet"`,
-		},
-		{
-			name: "invalid effective until",
-			mutate: func(manifest *networkManifest) {
-				manifest.Governance.EffectiveUntil = "not-a-timestamp"
-			},
-			wantErrSub: "invalid governance.effective_until",
-		},
-		{
-			name: "expired manifest",
-			mutate: func(manifest *networkManifest) {
-				manifest.Governance.EffectiveUntil = now.Add(-1 * time.Minute).Format(time.RFC3339)
-			},
-			wantErrSub: "expired at",
-		},
-		{
-			name: "unsupported source",
-			mutate: func(manifest *networkManifest) {
-				manifest.Policies.Source = "archive"
-				manifest.Policies.File = nil
-			},
-			wantErrSub: `uses unsupported policies.source "archive"`,
-		},
-		{
-			name: "signed file missing signature url",
-			mutate: func(manifest *networkManifest) {
-				manifest.Policies.File.SignatureURL = ""
-			},
-			wantErrSub: "requires policies.file.signature_url",
-		},
-		{
-			name: "signed bundle missing public key lookup",
-			mutate: func(manifest *networkManifest) {
-				manifest.Policies.Source = "bundle"
-				manifest.Policies.File = nil
-				manifest.Policies.Bundle = &networkManifestBundle{
-					ID:              "network-policy-bundle",
-					URL:             "https://example.com/policy.tar.gz",
-					PolicyQueryPath: "data.logistics.result",
-					Signed:          true,
-				}
-			},
-			wantErrSub: "requires policies.bundle.signing_public_key_lookup_url",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			manifest := validNetworkManifestForTest(expectedNetworkID, now)
-			if tt.mutate != nil {
-				tt.mutate(&manifest)
-			}
-
-			err := validateNetworkManifest(&manifest, expectedNetworkID, now)
-			if tt.wantErrSub == "" {
-				if err != nil {
-					t.Fatalf("expected success, got error: %v", err)
-				}
-				return
-			}
-			if err == nil {
-				t.Fatalf("expected error containing %q, got nil", tt.wantErrSub)
-			}
-			if !strings.Contains(err.Error(), tt.wantErrSub) {
-				t.Fatalf("expected error containing %q, got %v", tt.wantErrSub, err)
-			}
-		})
 	}
 }
